+++ test.lua
@@ -9,8 +9,8 @@ local fs = require 'bee.filesystem'
 ROOT = fs.path(rootPath)
 TEST = true
 DEVELOP = true
-FOOTPRINT = true
-TRACE = true
+--FOOTPRINT = true
+--TRACE = true
 LOGPATH  = LOGPATH  or (ROOT .. '/log')
 METAPATH = METAPATH or (ROOT .. '/meta')