+++ changelog.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## 2.0.2
+* `FIX` [#567](https://github.com/sumneko/lua-language-server/issues/567)
 * `FIX` [#570](https://github.com/sumneko/lua-language-server/issues/570)
 
 ## 2.0.1